requirements.txt
attrs
bs4
multipledispatch
numpy
pandas
retrying
seaborn
selenium
statsmodels
